Add user_id to tasks

+++ api/taskmasterexp/endpoints/tasks.py
@@ -3,6 +3,7 @@ from uuid import UUID
 from fastapi import APIRouter, status
 from fastapi.exceptions import HTTPException
 
+from taskmasterexp.auth import CurrentUser
 from taskmasterexp.database.dependencies import TaskManager
 from taskmasterexp.schemas.tasks import Task, TaskData, TaskResponse
 
@@ -10,41 +11,50 @@ router = APIRouter(prefix="/tasks", tags=["tasks"])
 
 
 @router.get("/", response_model=list[TaskResponse])
-async def list_tasks(manager: TaskManager, status: str = None):
-    filter = {}
+async def list_tasks(
+    current_user: CurrentUser, manager: TaskManager, status: str = None
+):
+    filter = {
+        "user_id": current_user.uuid,
+    }
 
     if status:
         filter["status"] = status
 
-    if filter:
-        tasks = await manager.list(filter)
-    else:
-        tasks = await manager.list()
+    tasks = await manager.list(filter)
     return tasks
 
 
 @router.post("/", status_code=status.HTTP_201_CREATED, response_model=TaskResponse)
-async def add_tasks(data: TaskData, manager: TaskManager):
-    task = Task(**data.dict())
+async def add_tasks(current_user: CurrentUser, manager: TaskManager, data: TaskData):
+    task = Task(user_id=current_user.uuid, **data.dict())
     task = await manager.save(task)
     return task
 
 
 @router.get("/{task_id}", response_model=TaskResponse)
-async def get_task(task_id: UUID, manager: TaskManager):
+async def get_task(current_user: CurrentUser, manager: TaskManager, task_id: UUID):
     task = await manager.get(task_id)
     if task is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND)
+
+    if task.user_id != current_user.uuid:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
 
     return task
 
 
 @router.patch("/{task_id}", response_model=TaskResponse)
-async def modify_task(task_id: UUID, data: TaskData, manager: TaskManager):
+async def modify_task(
+    current_user: CurrentUser, manager: TaskManager, task_id: UUID, data: TaskData
+):
     task = await manager.get(task_id)
 
     if task is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND)
+
+    if task.user_id != current_user.uuid:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
 
     task.update(data.dict(exclude_unset=True))
     await manager.save(task)
@@ -53,5 +63,11 @@ async def modify_task(task_id: UUID, data: TaskData, manager: TaskManager):
 
 
 @router.delete("/{task_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_task(task_id: UUID, manager: TaskManager):
-    await manager.delete(task_id)
+async def delete_task(current_user: CurrentUser, manager: TaskManager, task_id: UUID):
+    task = await manager.get(task_id)
+
+    if task:
+        if task.user_id != current_user.uuid:
+            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
+
+        await manager.delete(task_id)
